fixing issues

+++ shopping_cart/models.py
@@ -34,7 +34,7 @@ class Order(models.Model):
 class Payment(models.Model):
     stripe_charge_id = models.CharField(max_length=50)
     user = models.ForeignKey(User, on_delete=models.SET_NULL, blank=True, null=True)
-    amount = models.DecimalField(max_digits=100, decimal_places=2)
+    amount = models.IntegerField()
     timestamp = models.DateTimeField(auto_now_add=True)
